ADD: Refactored for model and controller packages.
