+++ app/src/main/java/io/ichi_go/ichigo/NewEventActivity.java
@@ -25,6 +25,8 @@ import java.io.OutputStreamWriter;
 import java.io.UnsupportedEncodingException;
 import java.io.Writer;
 
+import io.ichi_go.ichigo.data.model.Event;
+
 import static io.ichi_go.ichigo.R.id.event_description;
 import static io.ichi_go.ichigo.R.id.event_name;
