+++ app/src/main/java/io/ichi_go/ichigo/data/controller/EventManager.java
@@ -1,4 +1,4 @@
-package io.ichi_go.ichigo;
+package io.ichi_go.ichigo.data.controller;
 
 import android.util.Log;
 
@@ -6,7 +6,6 @@ import org.apache.http.HttpResponse;
 import org.apache.http.NameValuePair;
 import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.entity.UrlEncodedFormEntity;
-import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.entity.ByteArrayEntity;
 import org.apache.http.impl.client.DefaultHttpClient;
@@ -24,6 +23,8 @@ import java.io.OutputStreamWriter;
 import java.io.UnsupportedEncodingException;
 import java.io.Writer;
 import java.util.ArrayList;
+
+import io.ichi_go.ichigo.data.model.Event;
 
 /**
  * The event manager handles events and event accessories.
